refactor: canceled, transferproblems etc to correct components
